Removed erroneous 4-bit shift
Values are already 16-bits signed, with 12-bit data, so it can be returned as is. The datasheet isn't terribly clear here, but testing leads me to the conclusion that the code in this PR is the correct solution. The math makes sense for 12-bit values in a 16-bit signed wrapper, though not full 16-bit data. The DS states at +/-2g on the accel 1 lsb = 1mg, so at 12-bits we get 4096mg or +/-2g. The values seem the be returned already formed as 16-bit signed integers, though, so the right-shift is incorrect.